fix bugs ; add: create dep; delete dep; get one dep; get one Employee; also add data validate;

+++ src/main/java/com/example/demo/controller/DepartmentController.java
@@ -19,32 +19,34 @@ public class DepartmentController {
 
     public DepartmentController(DepartmentServiceImpl departmentService, Mapper mapper) {
         this.departmentService = departmentService;
-
         this.mapper = mapper;
     }
 
-    @GetMapping("/allDepWithUsers")
+    @GetMapping("/Departments")
     private List<DepartmentDto> getAllDepartmentsWithTheyUsers() {
         List<DepartmentDto> departmentDtoList = mapper.convertDepartmentListToDepartmentDtoList(departmentService.getAll());
         return departmentDtoList;
     }
 
-    @GetMapping(value = "/department/{id}")
-    public DepartmentDto getOneDepartment(@PathVariable UUID id) {
+    @GetMapping("/Departments/{id}")
+    public DepartmentDto getDepartmentByUUID(@PathVariable UUID id) {
         Department departmentById = departmentService.getDepartmentById(id);
         return mapper.toDepartmentDto(departmentById);
     }
 
-    @PostMapping("/createDepartment")
+    @PostMapping("/Departments")
     public DepartmentDto createDepartment(@RequestBody CreateDepartmentRequest createDepartmentRequest) {
-        Department department = departmentService.saveDepartment(mapper.toDepartmentFromCreateDepartmentRequest(createDepartmentRequest));
+        Department department = departmentService.saveDepartment(mapper.fromDtoWithEmployees(createDepartmentRequest));
         DepartmentDto toDepartmentDto = mapper.toDepartmentDto(department);
         return toDepartmentDto;
     }
 
-    @DeleteMapping(value = "/deleteDepartment/{id}")
+    @DeleteMapping("/Departments/{id}")
     public void deleteDepartmentDro(@PathVariable UUID id) {
         Department departmentById = departmentService.getDepartmentById(id);
+        if (departmentById == null){
+            throw new NullPointerException();
+        }
         departmentService.remove(departmentById);
     }
